image_kernel/pyproject.toml
[tool.poetry]
name = "image-kernel"
version = "0.1.0"
description = ""
authors = ["dfpaulino <dfpaulino@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tensorflow = "^2.17.0"
matplotlib = "^3.9.1"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
